Upload latest project after combine with teammate
I get the latest file from teammate's branch and combine the file together. After test the application is working, i uploaded the latest project to my branch so that my teammates able to get the latest file and commit to their branch. After all member commit the latest file to their branch then we only merge the branch so that there will no be so many conflict.